apps/human_review_streamlit_app/requirements.txt
streamlit
databricks-sql-connector
databricks-sdk>=0.73.0
pandas
python-dotenv
